refactor: centralize data serialization and growspace management within the coordinator and add a new serializer module.
