+++ custom_components/growspace_manager/coordinator.py
@@ -8,7 +8,6 @@ import uuid
 from datetime import date, datetime, timedelta
 from typing import TYPE_CHECKING, Any
 
-from homeassistant.const import STATE_ON, STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
@@ -18,12 +17,8 @@ if TYPE_CHECKING:
     from .irrigation_coordinator import IrrigationCoordinator
     from .vwc_irrigation_coordinator import VWCIrrigationCoordinator
 
-
-from .bayesian_data import VPD_STRESS_THRESHOLDS
 from .const import (
     DATE_FIELDS,
-    DEFAULT_FLOWER_EARLY_DAYS,
-    DEFAULT_VEG_EARLY_DAYS,
     DOMAIN,
     PlantStage,
 )
@@ -34,12 +29,11 @@ from .migration_manager import MigrationManager
 from .models import Growspace, GrowspaceCoordinatorData, GrowspaceEvent, Plant
 from .notification_manager import NotificationManager
 from .plant_lifecycle_manager import PlantLifecycleManager
+from .serializers import GrowspaceSerializer
 from .storage_manager import StorageManager
 from .strain_library import StrainLibrary
 from .utils import (
-    calculate_days_since,
     calculate_plant_stage,
-    days_to_week,
     format_date,
     generate_growspace_grid,
     parse_date_field,
@@ -91,6 +85,7 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
 
         self.hass = hass
         self._lock = asyncio.Lock()
+        self.serializer = GrowspaceSerializer(hass)
         self.growspaces: dict[str, Growspace] = {}
         self.plants: dict[str, Plant] = {}
         self.events: dict[str, list[GrowspaceEvent]] = {}
@@ -123,7 +118,7 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
             str, IrrigationCoordinator | VWCIrrigationCoordinator
         ] = {}
         self.dehumidifier_coordinators: dict[str, DehumidifierCoordinator] = {}
-        self.created_entity_ids: list[str] = []
+        self.created_entity_ids: list[tuple[str, str, str]] = []
 
         # Load data
         if data is None:
@@ -476,12 +471,21 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
         if self.data and isinstance(self.data, dict):
             recs = self.data.get("air_exchange_recommendations", {})
 
+        # Serialize growspaces for frontend and sensors
+        serialized_growspaces = {}
+        for growspace_id, growspace in self.growspaces.items():
+            plants = self.get_growspace_plants(growspace_id)
+            serialized_growspaces[growspace_id] = self.serializer.serialize_growspace(
+                growspace, plants
+            )
+
         self.data: GrowspaceCoordinatorData = {
             "growspaces": self.growspaces,
             "plants": self.plants,
             "notifications_sent": self._notifications_sent,
             "notifications_enabled": self._notifications_enabled,
             "_version": datetime.now().isoformat(),
+            "serialized_growspaces": serialized_growspaces,
             "air_exchange_recommendations": recs,
         }
 
@@ -572,6 +576,58 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
                 len(plants_to_remove),
             )
 
+    def _update_growspace_structure(
+        self, growspace: Growspace, kwargs: dict[str, Any], changes: list[str]
+    ) -> bool:
+        """Update growspace structure (dimensions)."""
+        updated = False
+        if "rows" in kwargs:
+            rows = int(kwargs["rows"])
+            if rows != growspace.rows:
+                changes.append(f"rows: {growspace.rows} -> {rows}")
+                growspace.rows = rows
+                updated = True
+
+        if "plants_per_row" in kwargs:
+            ppr = int(kwargs["plants_per_row"])
+            if ppr != growspace.plants_per_row:
+                changes.append(f"plants_per_row: {growspace.plants_per_row} -> {ppr}")
+                growspace.plants_per_row = ppr
+                updated = True
+        return updated
+
+    def _update_growspace_config(
+        self, growspace: Growspace, kwargs: dict[str, Any], changes: list[str]
+    ) -> bool:
+        """Update growspace configuration."""
+        updated = False
+        if "name" in kwargs:
+            name = kwargs["name"]
+            if name != growspace.name:
+                changes.append(f"name: {growspace.name} -> {name}")
+                growspace.name = name
+                updated = True
+
+        if "notification_target" in kwargs:
+            nt = kwargs["notification_target"]
+            nt = nt.strip() if nt else None
+            current = growspace.notification_target
+            if nt != current:
+                changes.append(f"notification_target: {current} -> {nt}")
+                growspace.notification_target = nt
+                updated = True
+
+        if "environment_config" in kwargs:
+            growspace.environment_config = kwargs["environment_config"]
+            changes.append("environment_config updated")
+            updated = True
+
+        if "irrigation_config" in kwargs:
+            growspace.irrigation_config = kwargs["irrigation_config"]
+            changes.append("irrigation_config updated")
+            updated = True
+        return updated
+
     async def async_update_growspace(
         self, growspace_id: str, **kwargs: dict[str, Any]
     ) -> None:
@@ -581,51 +637,16 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
                 raise ValueError(f"Growspace {growspace_id} not found")
 
             growspace = self.growspaces[growspace_id]
-            updated = False
-            changes = []
+            changes: list[str] = []
 
-            if "name" in kwargs:
-                name = kwargs["name"]
-                if name != growspace.name:
-                    changes.append(f"name: {growspace.name} -> {name}")
-                    growspace.name = name
-                    updated = True
+            # Update structure
+            struct_updated = self._update_growspace_structure(
+                growspace, kwargs, changes
+            )
+            # Update config
+            config_updated = self._update_growspace_config(growspace, kwargs, changes)
 
-            if "rows" in kwargs:
-                rows = int(kwargs["rows"])
-                if rows != growspace.rows:
-                    changes.append(f"rows: {growspace.rows} -> {rows}")
-                    growspace.rows = rows
-                    updated = True
-
-            if "plants_per_row" in kwargs:
-                ppr = int(kwargs["plants_per_row"])
-                if ppr != growspace.plants_per_row:
-                    changes.append(
-                        f"plants_per_row: {growspace.plants_per_row} -> {ppr}"
-                    )
-                    growspace.plants_per_row = ppr
-                    updated = True
-
-            if "notification_target" in kwargs:
-                nt = kwargs["notification_target"]
-                # Normalize
-                nt = nt.strip() if nt else None
-                current = growspace.notification_target
-                if nt != current:
-                    changes.append(f"notification_target: {current} -> {nt}")
-                    growspace.notification_target = nt
-                    updated = True
-
-            if "environment_config" in kwargs:
-                growspace.environment_config = kwargs["environment_config"]
-                changes.append("environment_config updated")
-                updated = True
-
-            if "irrigation_config" in kwargs:
-                growspace.irrigation_config = kwargs["irrigation_config"]
-                changes.append("irrigation_config updated")
-                updated = True
+            updated = struct_updated or config_updated
 
             if updated:
                 _LOGGER.info(
@@ -1093,253 +1114,7 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
         """Build the full JSON payload for a single growspace."""
         growspace = self.growspaces[growspace_id]
         plants = self.get_growspace_plants(growspace_id)
-
-        # Calculate max stage days
-        max_veg = max(
-            (calculate_days_since(p.veg_start) for p in plants if p.veg_start),
-            default=0,
-        )
-        max_flower = max(
-            (calculate_days_since(p.flower_start) for p in plants if p.flower_start),
-            default=0,
-        )
-        max_dry = max(
-            (calculate_days_since(p.dry_start) for p in plants if p.dry_start),
-            default=0,
-        )
-        max_cure = max(
-            (calculate_days_since(p.cure_start) for p in plants if p.cure_start),
-            default=0,
-        )
-
-        # Calculate weeks from days
-        veg_week = days_to_week(max_veg)
-        flower_week = days_to_week(max_flower)
-
-        biological_metrics = self._get_biological_metrics(
-            growspace, max_veg, max_flower, max_dry, max_cure
-        )
-
-        # Get irrigation settings
-        irrigation_options = growspace.irrigation_config
-        irrigation_strategy_dict = (
-            growspace.irrigation_strategy.to_dict()
-            if growspace.irrigation_strategy
-            else None
-        )
-
-        # Create grid representation (Detailed rich grid)
-        grid = self._generate_rich_plant_grid(growspace, plants)
-
-        # Determine growspace type
-        gs_type = "normal"
-        if growspace.id in ("mother", "clone", "dry", "cure"):
-            gs_type = growspace.id
-
-        # Build complete dict
-        data = {
-            "growspace_id": growspace.id,
-            "name": growspace.name,
-            "type": gs_type,
-            "rows": growspace.rows,
-            "plants_per_row": growspace.plants_per_row,
-            "total_plants": len(plants),
-            "notification_target": growspace.notification_target,
-            "max_veg_days": max_veg,
-            "max_flower_days": max_flower,
-            "veg_week": veg_week,
-            "flower_week": flower_week,
-            "max_stage_summary": f"Veg: {max_veg}d (W{veg_week}), Flower: {max_flower}d (W{flower_week})",
-            "irrigation_config": irrigation_options,
-            "irrigation_strategy": irrigation_strategy_dict,
-            "grid": grid,
-            **biological_metrics,
-        }
-
-        # Add environment attributes
-        data.update(self._get_environment_attributes(growspace))
-
-        return data
-
-    def _generate_rich_plant_grid(
-        self, growspace: Growspace, plants: list[Plant]
-    ) -> dict[str, dict[str, Any] | None]:
-        """Generate the detailed plant grid representation."""
-        grid: dict[str, dict[str, Any] | None] = {}
-        for row in range(1, int(growspace.rows) + 1):
-            for col in range(1, int(growspace.plants_per_row) + 1):
-                grid[f"position_{row}_{col}"] = None
-
-        # Fill grid with plants
-        for plant in plants:
-            row_i = int(plant.row)
-            col_i = int(plant.col)
-            position_key = f"position_{row_i}_{col_i}"
-            grid[position_key] = {
-                "plant_id": plant.plant_id,
-                "strain": plant.strain,
-                "phenotype": plant.phenotype,
-                # Days in stage
-                "seedling_days": self.calculate_days_in_stage(plant, "seedling"),
-                "mother_days": self.calculate_days_in_stage(plant, "mother"),
-                "clone_days": self.calculate_days_in_stage(plant, "clone"),
-                "veg_days": self.calculate_days_in_stage(plant, "veg"),
-                "flower_days": self.calculate_days_in_stage(plant, "flower"),
-                "dry_days": self.calculate_days_in_stage(plant, "dry"),
-                "cure_days": self.calculate_days_in_stage(plant, "cure"),
-                # Start dates
-                "seedling_start": format_date(plant.seedling_start),
-                "mother_start": format_date(plant.mother_start),
-                "clone_start": format_date(plant.clone_start),
-                "veg_start": format_date(plant.veg_start),
-                "flower_start": format_date(plant.flower_start),
-                "dry_start": format_date(plant.dry_start),
-                "cure_start": format_date(plant.cure_start),
-                # Location & Stage
-                "row": row_i,
-                "col": col_i,
-                "position": f"({row_i},{col_i})",
-                "stage": calculate_plant_stage(plant),
-            }
-        return grid
-
-    def _get_biological_metrics(
-        self,
-        growspace: Growspace,
-        max_veg: int,
-        max_flower: int,
-        max_dry: int,
-        max_cure: int,
-    ) -> dict[str, Any]:
-        """Calculate biological target metrics for the growspace."""
-        granular_stage = self._determine_granular_stage(
-            max_veg, max_flower, max_dry, max_cure
-        )
-        is_day = self._determine_is_day(growspace)
-        day_key = "day" if is_day else "night"
-
-        threshold_data = VPD_STRESS_THRESHOLDS.get(
-            granular_stage, VPD_STRESS_THRESHOLDS["veg_early"]
-        )
-        cycle_data = threshold_data.get(day_key, threshold_data["day"])
-
-        target_min, target_max = cycle_data.get("mild", (0.8, 1.2))
-        danger_min, danger_max = cycle_data.get("stress", (0.6, 1.4))
-
-        vpd_status = "unknown"
-        vpd_sensor_id = growspace.environment_config.get("vpd_sensor")
-
-        if vpd_sensor_id:
-            current_vpd = self._get_sensor_value(vpd_sensor_id)
-            if current_vpd is not None:
-                if current_vpd < danger_min or current_vpd > danger_max:
-                    vpd_status = "danger"
-                elif current_vpd < target_min or current_vpd > target_max:
-                    vpd_status = "warning"
-                else:
-                    vpd_status = "optimal"
-
-        return {
-            "granular_stage": granular_stage,
-            "is_day": is_day,
-            "vpd_target_min": target_min,
-            "vpd_target_max": target_max,
-            "vpd_danger_min": danger_min,
-            "vpd_danger_max": danger_max,
-            "vpd_status": vpd_status,
-        }
-
-    def _determine_granular_stage(
-        self, max_veg: int, max_flower: int, max_dry: int, max_cure: int
-    ) -> str:
-        """Determine granular growth stage based on days."""
-        if max_cure > 0:
-            return "cure"
-        if max_dry > 0:
-            return "dry"
-        if max_flower > 0:
-            if max_flower <= DEFAULT_FLOWER_EARLY_DAYS:
-                return "flower_early"
-            if max_flower <= (DEFAULT_FLOWER_EARLY_DAYS + 21):
-                return "flower_mid"
-            return "flower_late"
-        if max_veg > 0:
-            if max_veg > 0 and max_veg <= DEFAULT_VEG_EARLY_DAYS:
-                return "veg_early"
-            return "veg_late"
-        return "veg_early"
-
-    def _determine_is_day(self, growspace: Growspace) -> bool:
-        """Determine if it is currently day or night in the growspace."""
-        light_sensor = growspace.environment_config.get("light_sensor")
-        if light_sensor:
-            light_state = self.hass.states.get(light_sensor)
-            if light_state and light_state.state not in (
-                STATE_UNKNOWN,
-                STATE_UNAVAILABLE,
-            ):
-                if light_state.state == STATE_ON:
-                    return True
-                if light_state.state == "off":
-                    return False
-                try:
-                    return float(light_state.state) > 0
-                except (ValueError, TypeError):
-                    pass
-        return True
-
-    def _get_sensor_value(self, sensor_id: str | None) -> float | None:
-        """Safely get the numeric value from a sensor's state."""
-        if not sensor_id:
-            return None
-        state = self.hass.states.get(sensor_id)
-        if not state or state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN):
-            return None
-        try:
-            return float(state.state)
-        except (ValueError, TypeError):
-            return None
-
-    def _get_environment_attributes(self, growspace: Growspace) -> dict[str, Any]:
-        """Get environment-related attributes."""
-        attributes = {}
-        if growspace.environment_config:
-            env_config = growspace.environment_config
-            # Just copying minimal logic for now - or should we do full?
-            # The sensor had full logic including dehum, fan etc.
-            # Let's map the configured sensors.
-
-            # Map for aliasing or direct copy
-            # Frontend expects: exhaust_entity, humidifier_entity, dehumidifier_entity
-            # Config might have: exhaust_fan_entity, etc.
-
-            keys_to_map = {
-                "temperature_sensor": "temperature_sensor",
-                "humidity_sensor": "humidity_sensor",
-                "vpd_sensor": "vpd_sensor",
-                "co2_sensor": "co2_sensor",
-                "dehumidifier_entity": "dehumidifier_entity",
-                "humidifier_entity": "humidifier_entity",
-                "exhaust_entity": "exhaust_entity",
-                "exhaust_fan_entity": "exhaust_entity",
-                "circulation_fan_entity": "circulation_fan_entity",
-                "light_sensor": "light_sensor",
-                "control_dehumidifier": "dehumidifier_control_enabled",
-                "exhaust_sensor": "exhaust_sensor",
-                "humidifier_sensor": "humidifier_sensor",
-            }
-
-            for config_key, output_key in keys_to_map.items():
-                if val := env_config.get(config_key):
-                    attributes[output_key] = val
-
-                    if output_key == "dehumidifier_entity" and val:
-                        # Try to get state if possible, though frontend does this too
-                        state_obj = self.hass.states.get(val)
-                        if state_obj:
-                            attributes["dehumidifier_state"] = state_obj.state
-
-        return attributes
+        return self.serializer.serialize_growspace(growspace, plants)
 
     async def async_start_flowering(self, plant_id: str) -> Plant:
         """Transition a plant to the 'flower' stage, starting today."""
@@ -1520,30 +1295,6 @@ class GrowspaceCoordinator(DataUpdateCoordinator):
             for plant in self.plants.values()
             if plant.growspace_id == growspace_id
         ]
-
-    def calculate_days_in_stage(self, plant: Plant, stage: str) -> int:
-        """Calculate how many days a plant has been in a specific growth stage.
-
-        Args:
-            plant: The Plant object.
-            stage: The name of the stage (e.g., 'veg', 'flower').
-
-        Returns:
-            The number of days in the stage.
-        """
-        start_date = getattr(plant, f"{stage}_start", None)
-
-        end_date = None
-        if stage in {PlantStage.SEEDLING, PlantStage.CLONE}:
-            end_date = getattr(plant, "veg_start", None)
-        elif stage == PlantStage.VEG:
-            end_date = getattr(plant, "flower_start", None)
-        elif stage == PlantStage.FLOWER:
-            end_date = getattr(plant, "dry_start", None)
-        elif stage == PlantStage.DRY:
-            end_date = getattr(plant, "cure_start", None)
-
-        return self.calculate_days(start_date, end_date)
 
     def get_growspace_grid(self, growspace_id: str) -> list[list[str | None]]:
         """Generate a 2D grid representation of a growspace's plant layout.
