+++ custom_components/growspace_manager/manifest.json
@@ -12,12 +12,12 @@
     "intent"
   ],
   "documentation": "https://github.com/Venosta-web/growspace_manager",
-  "iot_class": "local_polling",
+  "iot_class": "local_push",
   "issue_tracker": "https://github.com/Venosta-web/growspace-manager/issues",
   "quality_scale": "silver",
   "requirements": [
     "aiosqlite",
     "Pillow"
   ],
-  "version": "0.2.5"
+  "version": "0.2.6"
 }
